backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LLM and AI
openai==1.10.0
anthropic==0.8.1
langchain==0.1.4
langchain-openai==0.0.5
tiktoken==0.5.2

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==4.0.1

# PDF generation
reportlab==4.0.9
weasyprint==60.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
aiofiles==23.2.1
websockets==12.0

# Logging
python-json-logger==2.0.7
